annotation_webapp/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
flask-cors>=4.0.0
Werkzeug>=3.0.0
